backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib==1.7.4
pydantic==2.5.2
python-dotenv==1.0.0
